Cleanup imports and move SPHINX_LT_?? constants to a centralized location
