+++ sphinx_automodapi/tests/example_module/__init__.py
@@ -1,3 +1,3 @@
-from .classes import *
-from .functions import *
-from .variables import *
+from .classes import *  # noqa
+from .functions import *  # noqa
+from .variables import *  # noqa
