+++ sphinx_automodapi/tests/helpers.py
@@ -5,15 +5,19 @@ import os
 import sys
 from copy import deepcopy
 
-from .test_cases import write_conf, SPHINX_LT_17, build_main
+from ..utils import SPHINX_LT_17
 from . import cython_testpackage  # noqa
 
-__all__ = ['run_sphinx_in_tmpdir']
+__all__ = ['build_main', 'write_conf', 'run_sphinx_in_tmpdir']
+
+if SPHINX_LT_17:
+    from sphinx import build_main
+else:
+    from sphinx.cmd.build import build_main
 
 intersphinx_mapping = {
     'python': ('https://docs.python.org/{0}/'.format(sys.version_info[0]), None)
     }
-
 
 DEFAULT_CONF = {'source_suffix': '.rst',
                 'master_doc': 'index',
@@ -25,6 +29,12 @@ DEFAULT_CONF = {'source_suffix': '.rst',
                 'automodapi_writereprocessed': True,
                 'automodapi_inheritance_diagram': True,
                 'automodsumm_writereprocessed': True}
+
+
+def write_conf(filename, conf):
+    with open(filename, 'w') as f:
+        for key, value in conf.items():
+            f.write("{0} = {1}\n".format(key, repr(conf[key])))
 
 
 def run_sphinx_in_tmpdir(tmpdir, additional_conf={}, expect_error=False):
